Refactor `request` interface
It doesn't make sense to provide a full `DebugProtocol.Request` directly
from the frontend, since the `seq` field is better set by `DebugClient`
in the middle-end. Since `DebugClient` already provides a convenience
function that takes a command and an `args` struct, provide the command,
vim_id, and an `args` struct to the middle-end from the frontend.