Robustify `CheckType`
When a TypeScript-to-nvim call fails due to an error, the Node.js log
output for the remote plugin will report the text of the error; this is
especially helpful when the error is an exception thrown by a
dapper.nvim type, since it returns much more useful information about
where the error occurred.

Refactor the `CheckType` function to always throw descriptive error.