Remove `is_start` from TypeScript DapperConfig
The `is_start` attribute was added based on the erroneous belief that
the Middleman might have to attach to a running debug adapter. In fact,
the Middleman will always *start* the debug adapter; it is the debug
adapter that can either launch a debugger, or attach to a running
debugger process.