Remove `is_start` from VimL types; tweak DapConf
See 83201908dd0b7cb1497c0bd9a96f7525536ba7d4 for an explanation of the
former.

Add a separate `filetype` argument to the DapperConfig constructor; this
would allow the user to register multiple debug adapters for a
particular filetype. (The `type` attribute given in VSCode `launch.json`
files might only say, e.g. `angular` or `node` to specify an AngularJS
or Node.js debug adapter, instead of giving the filetypes in which that
debug adapter might be useful.)