Fix variable range search logic
The previous implementation would not handle, e.g. the terminating
variable being one with one fewer levels of indentation than the current
variable; it would only stop at a Scope, or at another variable with the
same level of indentation.

Correct this by simply searching for a later line with less indentation.