Try to fix apparent race condition in VarsPrinter
Scope and variable callbacks may fire in the wrong order inside of
VariablesPrinter, which breaks things (e.g. a callback wants to print
the children of a variable, but that parent variable hasn't yet been
printed).

Try to address this by storing these "pending" prints, and firing them
once the parent on which they were waiting has been printed.