Fix bound arguments in delayed callback
The previous syntax simply provided PrintCollapsedChildren as a
callback; the provided arguments were instead treated as an error
handler by the call to Then. This would prevent the delayed callback
from working at all.

Add extra sleep statements in the test cases, to allow that delayed
callback to fire if necessary.