Gracefully abort pending requests post disconnect
If the running debug client was disconnected/terminated before a pending
request can be sent (in this case, the configurationDone request sent
from test case 'can start a mock debug adapter' in test_middleman.ts,
if it is not sent before the start of 'will terminate an already running
debug adapter'), the `this.dc.send` call will fail, because `dc` will be
an empty dict after the disconnection.

Check for this, and fail gracefully to avoid an unhandled Promise
rejection.