Further improve SetValue implementation
- Have SetVariable take an optional argument; if provided, use that as
the new variable value and don't prompt the user. (So that SetVariable
can be tested from test cases.)

- (Try to) update the Variable if the request succeeds. Currently, this
doesn't properly handle type changes or when the variable changes from
structured to unstructured.

- Where applicable, switch from `...Then(resolve, reject).Catch(reject)`
to `...Then(resolve).Catch(reject)` when the `reject` funcrefs are the
same. Does the same thing, but is more concise.